bug: small refactor (lol) to isolate the ts-ignore to its own file
