+++ aunty.config.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 
-const ADDITIONAL_ENTRY_POINTS = ['editor', 'standalone', 'polyfills', 'doc-block', 'illustrations', 'live', 'blanks'];
+const ADDITIONAL_ENTRY_POINTS = ['editor', 'standalone', 'doc-block', 'illustrations', 'live', 'blanks'];
 
 module.exports = {
   type: 'react',
