+++ public/blanks/index.html
@@ -1,37 +1,41 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Fill-in-the-blanks</title>
-    <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcsans/font.css" />
-    <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcserif/font.css" />
-    <style>
-      html {
-        font-family: ABCSans;
-      }
-      body {
-        margin: 0.5rem;
-      }
-      @media (max-width: 667px) {
-        .u-pull {
-          margin: 0 -0.5rem;
-        }
-      }
-    </style>
 
-    <script nomodule src="../polyfills.js"></script>
-    <script src="../blanks.js"></script>
-  </head>
-  <body>
-    <h1>Fill-in-the-blanks Examples</h1>
-    <h2>Live results</h2>
-    <div id="ecblanksLIVE" data-mount></div>
-    <h2>Live results (with state results)</h2>
-    <div id="ecblanksLIVESTATES" data-mount></div>
-    <h2>Empty</h2>
-    <div id="ecblanks" data-mount></div>
-    <h2>Custom</h2>
-    <div id="ecblanksALLOCATIONSr3ndnd3n2dnrdnr3d2ndn4r2nr3nrnd2ndnrdndr4n2d2nr2FOCUSESn51" data-mount></div>
-  </body>
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <title>Fill-in-the-blanks</title>
+  <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcsans/font.css" />
+  <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcserif/font.css" />
+  <style>
+    html {
+      font-family: ABCSans;
+    }
+
+    body {
+      margin: 0.5rem;
+    }
+
+    @media (max-width: 667px) {
+      .u-pull {
+        margin: 0 -0.5rem;
+      }
+    }
+  </style>
+
+  <script src="../blanks.js"></script>
+</head>
+
+<body>
+  <h1>Fill-in-the-blanks Examples</h1>
+  <h2>Live results</h2>
+  <div id="ecblanksLIVE" data-mount></div>
+  <h2>Live results (with state results)</h2>
+  <div id="ecblanksLIVESTATES" data-mount></div>
+  <h2>Empty</h2>
+  <div id="ecblanks" data-mount></div>
+  <h2>Custom</h2>
+  <div id="ecblanksALLOCATIONSr3ndnd3n2dnrdnr3d2ndn4r2nr3nrnd2ndnrdndr4n2d2nr2FOCUSESn51" data-mount></div>
+</body>
+
 </html>
