+++ public/doc-block/index.html
@@ -1,17 +1,19 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Google Docs Scrollyteller</title>
-    <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcsans/font.css" />
-    <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcserif/font.css" />
-    <link rel="stylesheet" href="../base.css" />
-    </style>
-    <script nomodule src="../polyfills.js"></script>
-    <script src="../doc-block.js"></script>
-  </head>
-  <body>
-    <div id="ecdb" data-mount></div>
-  </body>
+
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <title>Google Docs Scrollyteller</title>
+  <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcsans/font.css" />
+  <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcserif/font.css" />
+  <link rel="stylesheet" href="../base.css" />
+  </style>
+  <script src="../doc-block.js"></script>
+</head>
+
+<body>
+  <div id="ecdb" data-mount></div>
+</body>
+
 </html>
