+++ public/editor/index.html
@@ -1,16 +1,18 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Editor</title>
-    <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcsans/font.css" />
-    <!-- <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcserif/font.css" /> -->
-    </style>
-    <script nomodule src="../polyfills.js"></script>
-    <script src="../editor.js"></script>
-  </head>
-  <body>
-    <div id="eceditor" data-mount></div>
-  </body>
+
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <title>Editor</title>
+  <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcsans/font.css" />
+  <!-- <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcserif/font.css" /> -->
+  </style>
+  <script src="../editor.js"></script>
+</head>
+
+<body>
+  <div id="eceditor" data-mount></div>
+</body>
+
 </html>
