+++ public/illustration/index.html
@@ -1,13 +1,15 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Illustrations</title>
-    <script nomodule src="../polyfills.js"></script>
-    <script src="../illustrations.js"></script>
-  </head>
-  <body>
-    <div id="ecillustrations" data-mount></div>
-  </body>
+
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <title>Illustrations</title>
+  <script src="../illustrations.js"></script>
+</head>
+
+<body>
+  <div id="ecillustrations" data-mount></div>
+</body>
+
 </html>
