+++ public/live/index.html
@@ -1,15 +1,17 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Live Results</title>
-    <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcsans/font.css" />
-    <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcserif/font.css" />
-    <script nomodule src="../polyfills.js"></script>
-    <script src="../live.js"></script>
-  </head>
-  <body>
-    <div id="ecalllive" data-mount></div>
-  </body>
+
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <title>Live Results</title>
+  <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcsans/font.css" />
+  <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcserif/font.css" />
+  <script src="../live.js"></script>
+</head>
+
+<body>
+  <div id="ecalllive" data-mount></div>
+</body>
+
 </html>
