+++ public/standalone/index.html
@@ -1,30 +1,33 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Standalone</title>
-    <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcsans/font.css" />
-    <!-- <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcserif/font.css" /> -->
-    <style>
-      html {
-        font-family: ABCSans;
-      }
-      [data-mount] {
-        max-width: 960px;
-      }
-    </style>
-    <script nomodule src="../polyfills.js"></script>
-    <script src="../standalone.js"></script>
-  </head>
-  <body>
-    <h2>George's example of all allocation types</h2>
-    <div data-mount id="ecgraphicALLOCATIONSn3s3un2r2n3un11dn8sndn2en2dn3rdnunesn2FOCUSESn51RELATIVE2016"></div>
-    <h2>Safe</h2>
-    <div data-mount id="ecgraphicALLOCATIONSr3ndnd3n2dnrdnr3d2ndn4r2nr3nrnd2ndnrdndr4n2d2nr2FOCUSESn51"></div>
-    <h2>2016</h2>
-    <div data-mount id="ecgraphicALLOCATIONSr4d5r2dr2dr4d4r2dr9d5r2drdr5d3r3FOCUSESn51YEAR2016"></div>
-    <h2>2012</h2>
-    <div data-mount id="ecgraphicALLOCATIONSr4d6rd2rdr4d7r9d6rd3r5d4r2FOCUSESn51YEAR2012"></div>
-  </body>
+
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <title>Standalone</title>
+  <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcsans/font.css" />
+  <!-- <link rel="stylesheet" href="https://www.abc.net.au/res/fonts/abcserif/font.css" /> -->
+  <style>
+    html {
+      font-family: ABCSans;
+    }
+
+    [data-mount] {
+      max-width: 960px;
+    }
+  </style>
+  <script src="../standalone.js"></script>
+</head>
+
+<body>
+  <h2>George's example of all allocation types</h2>
+  <div data-mount id="ecgraphicALLOCATIONSn3s3un2r2n3un11dn8sndn2en2dn3rdnunesn2FOCUSESn51RELATIVE2016"></div>
+  <h2>Safe</h2>
+  <div data-mount id="ecgraphicALLOCATIONSr3ndnd3n2dnrdnr3d2ndn4r2nr3nrnd2ndnrdndr4n2d2nr2FOCUSESn51"></div>
+  <h2>2016</h2>
+  <div data-mount id="ecgraphicALLOCATIONSr4d5r2dr2dr4d4r2dr9d5r2drdr5d3r3FOCUSESn51YEAR2016"></div>
+  <h2>2012</h2>
+  <div data-mount id="ecgraphicALLOCATIONSr4d6rd2rdr4d7r9d6rd3r5d4r2FOCUSESn51YEAR2012"></div>
+</body>
+
 </html>
