--- src/polyfills.ts
@@ -1,1 +1,0 @@
-import '@next/polyfill-nomodule';
